requirements.txt
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
flask-swagger-ui==4.11.1
googlemaps==4.10.0
idna==3.7
importlib_metadata==7.2.1
itsdangerous==2.2.0
Jinja2==3.1.4
mailjet-rest==1.3.4
MarkupSafe==2.1.5
mongoengine==0.28.2
PyJWT==2.8.0
pymongo==4.7.3
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.19.2
flask_sqlalchemy
flask_migrate